+++ confidence.py
@@ -10,12 +10,35 @@ import yaml
 
 
 class ConfigurationError(KeyError):
-    """
-    `KeyError` raised when merge conflicts are detected during `.Configuration`
-    construction (see `.Configuration.__init__`) or retrieving an unavailable
-    configured value when no default is supplied (see `.Configuration.get`).
-    """
     pass
+
+
+class MergeConflictError(ConfigurationError):
+    """
+    Error raised during loading configuration sources that conflict internally.
+    """
+    def __init__(self, *args, key):
+        super().__init__(*args)
+        self.conflict = key
+
+
+class NotConfiguredError(ConfigurationError):
+    """
+    Error raised when a requested configuration key is unavailable and no
+    default / fallback value is provided.
+    """
+    def __init__(self, *args, key):
+        super().__init__(*args)
+        self.key = key
+
+
+class ConfiguredReferenceError(ConfigurationError):
+    """
+    Error raised a referenced configuration key is unavailable.
+    """
+    def __init__(self, *args, key):
+        super().__init__(*args)
+        self.key = key
 
 
 class _Conflict(IntEnum):
@@ -46,7 +69,8 @@ def _merge(left, right, path=None, conflict=_Conflict.error):
             elif left[key] != right[key]:
                 if conflict is _Conflict.error:
                     # not both dicts we could merge, but also not the same, this doesn't work
-                    raise ConfigurationError('merge conflict at {}'.format('.'.join(path + [key])))
+                    conflict_path = '.'.join(path + [key])
+                    raise MergeConflictError('merge conflict at {}'.format(conflict_path), key=conflict_path)
                 else:
                     # overwrite left value with right value
                     left[key] = right[key]
@@ -111,16 +135,19 @@ class Configuration(Mapping):
     or attributes.
     """
 
+    # match a reference as ${key.to.be.resolved}
+    _reference_pattern = re.compile(r'\${(?P<path>[^}]+?)}')
+
     def __init__(self, *sources, separator='.'):
         """
         Create a new `.Configuration`, based on one or multiple source mappings.
 
         :param sources: source mappings to base this `.Configuration` on,
             ordered from least to most significant
-        :param separator: the character (sequence) to use as the separator
-            between keys
+        :param separator: the character(s) to use as the separator between keys
         """
         self._separator = separator
+        self._root = self
 
         self._source = {}
         for source in sources:
@@ -128,7 +155,55 @@ class Configuration(Mapping):
                 # merge values from source into self._source, overwriting any corresponding keys
                 _merge(self._source, _split_keys(source, separator=self._separator), conflict=_Conflict.overwrite)
 
-    def get(self, path, default=_NoDefault, as_type=None):
+    def _resolve(self, value):
+        match = self._reference_pattern.search(value)
+        references = set()
+        try:
+            while match:
+                path = match.group('path')
+                if path in references:
+                    raise ConfiguredReferenceError(
+                        'cannot resolve recursive reference {path}'.format(path=path),
+                        key=path
+                    )
+
+                # avoid resolving references recursively (breaks reference tracking)
+                reference = self._root.get(path, resolve_references=False)
+
+                if match.span(0) != (0, len(value)):
+                    # matched a reference inside of another value (template)
+                    if isinstance(reference, Configuration):
+                        raise ConfiguredReferenceError(
+                            'cannot insert namespace at {path} into referring value'.format(path=path),
+                            key=path
+                        )
+
+                    # render the template containing the referenced value
+                    value = '{start}{reference}{end}'.format(
+                        start=value[:match.start(0)],
+                        reference=reference,
+                        end=value[match.end(0):]
+                    )
+                else:
+                    # value is only a reference, avoid rendering a template (keep referenced value type)
+                    value = reference
+
+                # track that we've seen path
+                references.add(path)
+                # either keep finding references or stop resolving and return value
+                if isinstance(value, str):
+                    match = self._reference_pattern.search(value)
+                else:
+                    match = None
+
+            return value
+        except NotConfiguredError as e:
+            raise ConfiguredReferenceError(
+                'unable to resolve referenced key {reference}'.format(reference=match.group('path')),
+                key=e.key
+            ) from e
+
+    def get(self, path, default=_NoDefault, as_type=None, resolve_references=True):
         """
         Gets a value for the specified path.
 
@@ -140,10 +215,11 @@ class Configuration(Mapping):
         :param as_type: an optional callable to apply to the value found for
             the supplied path (possibly raising exceptions of its own if the
             value can not be coerced to the expected type)
+        :param resolve_references: whether to resolve references in values
         :return: the value associated with the supplied configuration key, if
             available, or a supplied default value if the key was not found
         :raises ConfigurationError: when no value was found for *path* and
-            *default* was not provided
+            *default* was not provided or a reference could not be resolved
         """
         value = self._source
         steps_taken = []
@@ -158,14 +234,23 @@ class Configuration(Mapping):
             elif isinstance(value, Mapping):
                 namespace = Configuration()
                 namespace._source = value
+                # carry the root object from namespace to namespace, references are always resolved from root
+                namespace._root = self._root
                 return namespace
+            elif resolve_references and isinstance(value, str):
+                # only resolve references in str-type values (the only way they can be expressed)
+                return self._resolve(value)
             else:
                 return value
+        except ConfiguredReferenceError:
+            # also a KeyError, but this one should bubble to caller
+            raise
         except KeyError:
             if default is not _NoDefault:
                 return default
             else:
-                raise ConfigurationError('no configuration for key {}'.format(self._separator.join(steps_taken)))
+                missing_key = self._separator.join(steps_taken)
+                raise NotConfiguredError('no configuration for key {}'.format(missing_key), key=missing_key)
 
     def __getattr__(self, attr):
         """
@@ -224,7 +309,7 @@ class NotConfigured(Configuration):
 
 
 # overwrite NotConfigured as an instance of itself, a Configuration instance without any values
-NotConfigured = NotConfigured({})
+NotConfigured = NotConfigured()
 
 
 def load(*fps):